REF: verficationService를 읽기 전용 서비스로 분리
