+++ src/main/java/com/planup/planup/domain/verification/service/PhotoVerificationService.java
@@ -64,26 +64,4 @@ public class PhotoVerificationService implements VerificationService {
 
         photoVerificationRepository.delete(verification);
     }
-
-    @Transactional(readOnly = true)
-    public List<PhotoVerification> getPhotoVerificationListByUserAndDateBetween(UserGoal userGoal, LocalDateTime start, LocalDateTime end) {
-        return photoVerificationRepository.findAllByUserGoalAndCreatedAtBetweenOrderByCreatedAt(userGoal,start,end);
-    }
-
-    @Override
-    public Map<LocalDate, Integer> calculateVerification(UserGoal userGoal, LocalDateTime startDate, LocalDateTime endDate) {
-        List<PhotoVerification> verifications = getPhotoVerificationListByUserAndDateBetween(userGoal, startDate, endDate);
-
-        Map<LocalDate, Integer> dailyCount = new HashMap<>();
-        // 날짜별 인증 수 카운팅
-        for (PhotoVerification photoVerification : verifications) {
-            LocalDate date = photoVerification.getCreatedAt().toLocalDate();
-
-            int photoCount = photoVerification.getPhotoImgs() != null ? photoVerification.getPhotoImgs().size() : 0;
-
-            //기존에 데이터가 있으면 불러와서 더한다.
-            dailyCount.put(date, dailyCount.getOrDefault(date, 0) + photoCount);
-        }
-        return dailyCount;
-    }
 }
