+++ src/main/java/com/planup/planup/domain/verification/service/VerificationService.java
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 import java.util.Map;
 
 public interface VerificationService {
-    @Transactional
-    Map<LocalDate, Integer> calculateVerification(UserGoal userGoal, LocalDateTime startDate, LocalDateTime endDate);
+//    @Transactional
+//    Map<LocalDate, Integer> calculateVerification(UserGoal userGoal, LocalDateTime startDate, LocalDateTime endDate);
 }
